refactored code for sanitizing urls into separate modules
